refactor: upsertItemInListCache を src/lib/itemCache.ts へ移動
types/item.ts はZod/型定義のみのモジュールのため、ランタイムロジックを
専用ファイル src/lib/itemCache.ts へ移動し責務を分離。

あわせて Array.isArray 後の中間変数 `const list` を除去し
インラインキャストに整理（Copilot レビュー指摘対応）。
