+++ src/types/item.ts
@@ -90,14 +90,6 @@ export type UserSettings = z.infer<typeof userSettingsSchema>;
 
 export type ExpiryStatus = "expired" | "expiring-soon" | "ok" | "unknown";
 
-export const upsertItemInListCache = (old: unknown, incoming: Item): Item[] | undefined => {
-  if (!Array.isArray(old)) return undefined;
-  const list = old as Item[];
-  const index = list.findIndex((item) => item.id === incoming.id);
-  if (index === -1) return [incoming, ...list];
-  return list.map((item) => (item.id === incoming.id ? incoming : item));
-};
-
 export const DEFAULT_EXPIRY_WARNING_DAYS = 3;
 
 export const CONTENT_UNITS = ["個", "枚", "本", "袋", "mL", "L", "g", "kg"] as const;
